added compare projects for private and public cloud
